+++ CHANGELOG.md
@@ -93,3 +93,4 @@
 ## 3.0.6
 
 - Merge library
+- Complete coverage